Lo puse más bonito de lo que ya estaba y completo en español
